requirements.txt
h5py==2.10.0
matplotlib==3.3.4
numpy==1.20.1
pandas==1.2.4
Pillow==8.4.0
scikit_image==0.18.1
streamlit==1.1.0
torch==1.10.1
torchvision==0.11.2
